change specobjid to decimal type

+++ Sciserver_webService/ToolsSearch/ObjectInfo.cs
@@ -9,7 +9,7 @@ namespace Sciserver_webService.ToolsSearch
     {
         public long? id = null;
         public string apid;
-        public long? specId = null;
+        public decimal? specId = null;
 
         public string objId = null;
         public string specObjId = null;
